+++ src/components/molecules/FooterContent.js
@@ -1,0 +1,5 @@
+function FooterContent() {
+  return <span>fgosc project</span>;
+}
+
+export default FooterContent;
