+++ src/components/organisms/FooterComponent.js
@@ -1,0 +1,7 @@
+import FooterContent from "../molecules/FooterContent";
+
+function FooterComponent() {
+  return <FooterContent />;
+}
+
+export default FooterComponent;
