+++ src/components/organisms/HeaderComponent.js
@@ -1,0 +1,7 @@
+import { Heading } from "@chakra-ui/react";
+
+function HeaderComponent() {
+  return <Heading>FGO 周回報告</Heading>;
+}
+
+export default HeaderComponent;
