+++ src/contexts/UserAttributesContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const UserAttributesContext = createContext();
+
+export default UserAttributesContext;
